Working imports.

+++ app/apps/imports/parsers.py
@@ -10,6 +10,7 @@ from django.utils.translation import gettext as _
 from bs4 import BeautifulSoup
 
 from core.models import *
+from versioning.models import NoChangeException
 
 
 XML_EXTENSIONS = ['xml', 'alto', 'abbyy']
@@ -28,14 +29,16 @@ class XMLParser():
     SCHEMA_ALTO = 'alto'
     DEFAULT_NAME = _("XML Import")
     
-    def __init__(self, root, name=None):
+    def __init__(self, root, name=None, override=True):
         self.root = root
         self.name = name or self.DEFAULT_NAME
-        self.pages = self.get_pages()
-    
-    def get_pages(self):
-        raise NotImplemented
-    
+        self.override = override
+        self.namespace = self.root.nsmap[None]
+        self.pages = self.find(self.root, self.TAGS['page'])
+
+    def find(self, parent, tag):
+        return parent.findall('.//{%s}%s' % (self.namespace, tag))
+        
     def make_line(self, line_tag):
         pass
     
@@ -65,31 +68,52 @@ class XMLParser():
         trans, created = Transcription.objects.get_or_create(
             document=document,
             name=self.name)
-        
         try:
             for index, page in enumerate(self.pages):
                 if index < start_at:
                     continue
-                with transaction.atomic():  
+                with transaction.atomic():
                     part = parts[index]
-                    part.blocks.all().delete()
-                    part.lines.all().delete()
-                    for block in page.findall(self.TAGS['block']):
-                        b = Block.objects.create(
-                            document_part=part,
-                            box=self.block_bbox(block))
-                        
-                        for line in block.findall(self.TAGS['line']):
-                            l = Line.objects.create(
-                                document_part=part,
-                                block=b,
-                                box=self.line_bbox(line))
+                    if self.override:
+                        part.blocks.all().delete()
+                        part.lines.all().delete()
+                    for block in self.find(page, self.TAGS['block']):
+                        # Note: don't use get_or_create to avoid a update query
+                        attrs = {'document_part': part,
+                                 'external_id': block.get('ID')}
+                        try:
+                            b = Block.objects.get(**attrs)
+                        except Block.DoesNotExist:
+                            b = Block(**attrs)
+                        b.box = self.block_bbox(block)
+                        b.save()
+                        for line in self.find(block, self.TAGS['line']):
+                            attrs = {'document_part': part,
+                                     'block': b,
+                                     'external_id': line.get('ID')}
+                            try:
+                                l = Line.objects.get(**attrs)
+                            except Line.DoesNotExist:
+                                l = Line(**attrs)
+                            l.box = self.line_bbox(line)
+                            l.save()
                             content = self.make_line(line)
                             if content:
-                                lt = LineTranscription.objects.create(
-                                    transcription=trans,
-                                    line=l,
-                                    content=content)
+                                attrs = {'transcription': trans, 'line':l}
+                                try:
+                                    lt = LineTranscription.objects.get(**attrs)
+                                except LineTranscription.DoesNotExist:
+                                    lt = LineTranscription(version_source='import',
+                                                           version_author=self.name,
+                                                           **attrs)
+                                else:
+                                    try:
+                                        lt.new_version()  # save current content in history
+                                    except NoChangeException:
+                                        pass
+                                lt.content = content
+                                lt.save()
+                            
                             # TODO: store glyphs too
                 self.post_process(part)
                 yield part
@@ -105,28 +129,17 @@ class AltoParser(XMLParser):
     }
     DEFAULT_NAME = _("ALTO Import")
     SCHEMA = 'http://www.loc.gov/standards/alto/v4/alto-4-0.xsd'
-
-    def get_pages(self):
-        try:
-            from IPython import embed
-            embed()
-            return self.root.find('Layout').findall(self.TAGS['page'])
-        except AttributeError as e:
-            raise ParseError
     
     def bbox(self, tag):
         return (
-            int(tag.attrs['HPOS']),
-            int(tag.attrs['VPOS']),
-            int(tag.attrs['HPOS']) + int(tag.attrs['WIDTH']),
-            int(tag.attrs['VPOS']) + int(tag.attrs['HEIGHT']),
+            int(tag.get('HPOS')),
+            int(tag.get('VPOS')),
+            int(tag.get('HPOS')) + int(tag.get('WIDTH')),
+            int(tag.get('VPOS')) + int(tag.get('HEIGHT')),
         )
     
     def make_line(self, line_tag):
-        content = ''
-        for string in line_tag.findall('String'):
-            content += (string.attrs.get('CONTENT') + ' ')
-        return content
+        return ' '.join([e.get('CONTENT') for e in self.find(line_tag, 'String')])
 
     def validate(self):
         try:
@@ -137,10 +150,8 @@ class AltoParser(XMLParser):
             schema_root = etree.XML(response.content)
             xmlschema = etree.XMLSchema(schema_root)
             try:
-                xmlschema.assertValid(self.soup)
+                xmlschema.assertValid(self.root)
             except (AttributeError, etree.DocumentInvalid, etree.XMLSyntaxError) as e:
-                if settings.DEBUG:
-                    print(e)
                 raise ParseError("Document didn't validate. %s" % e.args[0])
 
 
@@ -149,25 +160,18 @@ class AbbyyParser(XMLParser):
     DEFAULT_NAME = _("ABBYY Import")
     SCHEMA = ''
     
-    def get_pages(self):
-        try:
-            return self.root.findall(self.TAGS['page'])
-        except AttributeError as e:
-            raise ParseError
-    
     def line_bbox(self, tag):
         return (
-            int(tag.attrs['l']),
-            int(tag.attrs['t']),
-            int(tag.attrs['r']),
-            int(tag.attrs['b'])
-        )
+            int(tag.get('l')),
+            int(tag.get('t')),
+            int(tag.get('r')),
+            int(tag.get('b')))
     
     def block_bbox(self, tag):
         return (0,0,0,0)
     
     def make_line(self, line_tag):
-        content = ''.join(map(lambda a: a.text, line_tag.findall('charParams')))
+        content = ''.join([e.text for e in self.find(line_tag, 'charParams')])
         return content
 
     def post_process(self, part):
@@ -236,22 +240,25 @@ class IIIFManifesParser():
 
         
 
-def make_parser(file_handler):
+def make_parser(file_handler, name=None, override=True):
     # TODO: not great to rely on extension
     ext = os.path.splitext(file_handler.name)[1][1:]
     if ext in XML_EXTENSIONS:
-        root = etree.parse(file_handler).getroot()
+        try:
+            root = etree.parse(file_handler).getroot()
+        except etree.XMLSyntaxError as e:
+            raise ParseError(e.msg)
         try:
             schema = root.nsmap[None]
         except KeyError:
             raise ParseError("Couldn't determine xml schema, xmlns attribute missing on root element.")
         if 'abbyy' in schema:  # Not super robust
-            return AbbyyParser(root)
+            return AbbyyParser(root, name=name, override=override)
         elif 'alto' in schema:
-            return AltoParser(root)
+            return AltoParser(root, name=name, override=override)
         else:
             raise ParseError("Couldn't determine xml schema, abbyy or alto, check the content of the root tag.")
     elif ext == 'json':
-        return IIIFManifesParser(file_handler)
+        return IIIFManifesParser(file_handler, name=name)
     else:
         raise ValueError("Invalid extension for the file to be parsed %s." % file_handler.name)
